requirements.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
brotli==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
brotlicffi==1.1.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation == "PyPy"
build==1.0.3 ; python_version >= "3.11" and python_version < "4.0"
cachecontrol[filecache]==0.13.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or os_name == "nt" or platform_system == "Windows")
crashtest==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
cryptography==42.0.2 ; python_version >= "3.11" and python_version < "4.0"
distlib==0.3.8 ; python_version >= "3.11" and python_version < "4.0"
docker==7.0.0 ; python_version >= "3.11" and python_version < "4.0"
dulwich==0.21.7 ; python_version >= "3.11" and python_version < "4.0"
fastjsonschema==2.19.1 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.13.1 ; python_version >= "3.11" and python_version < "4.0"
flask==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==7.0.1 ; python_version >= "3.11" and python_version < "3.12"
inflate64==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
installer==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jaraco-classes==3.3.1 ; python_version >= "3.11" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
jinja2==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
kaitaistruct==0.10 ; python_version >= "3.11" and python_version < "4.0"
keyring==24.3.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4.0"
more-itertools==10.2.0 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
multivolumefile==0.2.3 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "4.0"
opencv-python==4.9.0.80 ; python_version >= "3.11" and python_version < "4.0"
outcome==1.3.0.post0 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
pkginfo==1.9.6 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==3.11.0 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
poetry-core==1.8.1 ; python_version >= "3.11" and python_version < "4.0"
poetry-plugin-export==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
poetry==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
psutil==5.9.8 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "cygwin"
ptyprocess==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
py7zr==0.20.8 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
pybcj==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pycryptodomex==3.20.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.16.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
pydivert==2.1.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
pyopenssl==24.1.0 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
pyppmd==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pyproject-hooks==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.0.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
python-magic==0.4.27 ; python_version >= "3.11" and python_version < "4.0"
pywin32-ctypes==0.2.2 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
pywin32==306 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
pyzstd==0.15.9 ; python_version >= "3.11" and python_version < "4.0"
rapidfuzz==3.6.1 ; python_version >= "3.11" and python_version < "4.0"
rarfile==4.1 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
ruff==0.2.2 ; python_version >= "3.11" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
selenium-wire==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
selenium==4.19.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==69.5.1 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
tdqm==0.0.1 ; python_version >= "3.11" and python_version < "4.0"
texttable==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.12.3 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.2 ; python_version >= "3.11" and python_version < "4.0"
trio-websocket==0.11.1 ; python_version >= "3.11" and python_version < "4.0"
trio==0.25.0 ; python_version >= "3.11" and python_version < "4.0"
trove-classifiers==2024.1.31 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
undetected-chromedriver==3.5.5 ; python_version >= "3.11" and python_version < "4.0"
unrar==0.4 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3[socks]==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
virtualenv==20.25.0 ; python_version >= "3.11" and python_version < "4.0"
webdriver-manager==4.0.1 ; python_version >= "3.11" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
xattr==0.10.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.17.0 ; python_version >= "3.11" and python_version < "3.12"
zstandard==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
